requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
camelot-py==1.0.9
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
dataclasses-json==0.6.7
emoji==2.15.0
et_xmlfile==2.0.0
Faker==40.1.2
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.5
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langchain-unstructured==1.0.1
langdetect==1.0.9
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.7
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.4.2
olefile==0.47
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
orjson==3.11.6
ormsgpack==1.12.2
packaging==25.0
pandas==3.0.0
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pi_heif==1.2.0
pillow==12.1.0
pillow_heif==1.2.0
poppler-utils==0.1.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdf==5.9.0
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.2
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.31
unstructured-client==0.42.8
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.0
webencodings==0.5.1
wrapt==2.1.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
